+++ src/ui/components/tests/layer-membership-list-test.js
@@ -33,7 +33,7 @@ describe('layer-membership-list', function() {
     });
     query.isFiring = false;
     query.data = [];
-    for (i = 0; i < 100; i++) {
+    for (i = 0; i < 35; i++) {
 
         var ident = new Layer.Core.Identity({
               userId: 'user' + i,