+++ src/ui/mixins/tests/list-load-indicator-test.js
@@ -33,7 +33,7 @@ describe("List Load Mixin", function() {
       predicate: 'conversation.id = "' + conversation.id + '"'
     });
     query.isFiring = false;
-    for (i = 0; i < 100; i++) {
+    for (i = 0; i < 35; i++) {
       query.data.push(conversation.createMessage("m " + i).send());
     }